// server/type.ts
export type IUser = {
  id: string;
  name: string;
  email: string;
  createdAt: Date;
  updatedAt: Date;
};

export enum NotificationType {
  AddedParticipant = "added_participant",
  AddedCollaborator = "added_collaborator",
  TaskReminder = "task_reminder",
  WelcomeMessage = "welcome_message",
  LoginAlert = "login_alert",
}

export type AddedParticipantMetadata = {
  task_link: string;
};

export type INotification = {
  id: string;
  type: NotificationType;
  title: string;
  message: string;
  is_read: boolean;
  user_id: string;
  metadata?: any;
  createdAt: Date;
  updatedAt: Date;
};

export type _IUser = {
  id: string;
  name: string;
  email?: string;
};
// ...
export type INote = {
  id: string;
  title: string;
  content: string;
  owner_id: string;
  self_destroy_time?: Date;
  category_name?: string;
  tags?: string[];
  last_indexed?: Date;
  createdAt: Date;
  updatedAt: Date;
};
// ...
export type ITask = {
  id: string;
  note_id: string;
  user_id: string;
  name: string;
  date: Date;
  reminder: Date;
  duration?: string;
  calendar_id?: string;
  location?: any;
  createdAt: Date;
  updatedAt: Date;
};
// ...
export type IOtpSession = {
  expiry: string | Date;
  email: string;
  auth_code_hash: string;
};

export type IAuthSession = {
  expiry: Date | string;
  ip_address: string | any;
  user_agent: string;
  user_id: string;
  socket_auth_hash: string;
};
export type IncomingNote = {
  self_destroy_time: string;
  content: string;
  title: string;
};
// ...
export enum ErrorCodes {
  VALIDATION_ERROR = "VALIDATION_ERROR",
  RESOURCE_NOT_FOUND = "RESOURCE_NOT_FOUND",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  CONFLICT = "CONFLICT",
  PAYMENT_REQUIRED = "PAYMENT_REQUIRED",
  INSUFFICIENT_FUNDS = "INSUFFICIENT_FUNDS",
  SUBSCRIPTION_EXPIRED = "SUBSCRIPTION_EXPIRED",
  PAYMENT_VERIFICATION_FAILED = "PAYMENT_VERIFICATION_FAILED",
}

// Email Drafting Job
type IEmailDraftJob = {
  job_type: "email_draft";
  payload: {
    recipient: string;
    subject: string;
    body: string;
    draft_id?: string; // Optional if saving draft
  };
};

export type ISummaryResponse = {
  summary: string;
  summary_id: string;
};